+++ Model/UserInterface.php
@@ -14,11 +14,16 @@ namespace FOS\UserBundle\Model;
 use Symfony\Component\Security\Core\User\UserInterface as BaseUserInterface;
 
 /**
+ * Implementations of that interface must be serializable. The mechanism
+ * being used to support serialization is up for the implementation.
+ * The serialized data must contain all the fields involved in comparing
+ * the user in the security component.
+ *
  * @author Thibault Duplessis <thibault.duplessis@gmail.com>
  * @author Johannes M. Schmitt <schmittjoh@gmail.com>
  * @author Julian Finkler <julian@developer-heaven.de>
  */
-interface UserInterface extends BaseUserInterface, \Serializable
+interface UserInterface extends BaseUserInterface
 {
     public const ROLE_DEFAULT = 'ROLE_USER';