+++ DependencyInjection/Compiler/CheckForSessionPass.php
@@ -30,8 +30,8 @@ class CheckForSessionPass implements CompilerPassInterface
      */
     public function process(ContainerBuilder $container)
     {
-        if ($container->has('fos_user.session') && !$container->has('session')) {
-            $message = 'FOSUserBundle requires the "session" service to be available.';
+        if ($container->has('fos_user.session') && !$container->has('session.storage.factory') && !$container->has('session')) {
+            $message = 'FOSUserBundle requires the "session" to be available for the enabled features.';
 
             if (class_exists(Recipe::class)) {
                 $message .= ' Uncomment the "session" section in "config/packages/framework.yaml" to activate it.';