+++ Tests/EventListener/AuthenticationListenerTest.php
@@ -30,7 +30,7 @@ class AuthenticationListenerTest extends TestCase
     /** @var AuthenticationListener */
     private $listener;
 
-    public function setUp()
+    protected function setUp(): void
     {
         $user = $this->getMockBuilder('FOS\UserBundle\Model\UserInterface')->getMock();