+++ Tests/EventListener/FlashListenerTest.php
@@ -14,7 +14,10 @@ namespace FOS\UserBundle\Tests\EventListener;
 use FOS\UserBundle\EventListener\FlashListener;
 use FOS\UserBundle\FOSUserEvents;
 use PHPUnit\Framework\TestCase;
+use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\HttpFoundation\RequestStack;
 use Symfony\Contracts\EventDispatcher\Event;
+use Symfony\Contracts\Translation\TranslatorInterface;
 
 class FlashListenerTest extends TestCase
 {
@@ -24,7 +27,7 @@ class FlashListenerTest extends TestCase
     /** @var FlashListener */
     private $listener;
 
-    public function setUp()
+    protected function setUp(): void
     {
         $this->event = new Event();
 
@@ -36,9 +39,15 @@ class FlashListenerTest extends TestCase
             ->method('getFlashBag')
             ->willReturn($flashBag);
 
-        $translator = $this->getMockBuilder('Symfony\Component\Translation\TranslatorInterface')->getMock();
+        $request = new Request();
+        $request->setSession($session);
 
-        $this->listener = new FlashListener($session, $translator);
+        $requestStack = new RequestStack();
+        $requestStack->push($request);
+
+        $translator = $this->getMockBuilder(TranslatorInterface::class)->getMock();
+
+        $this->listener = new FlashListener($requestStack, $translator);
     }
 
     public function testAddSuccessFlash()