+++ Tests/Form/Type/ProfileFormTypeTest.php
@@ -33,10 +33,7 @@ class ProfileFormTypeTest extends ValidatorExtensionTypeTestCase
         $this->assertSame('john@doe.com', $user->getEmail());
     }
 
-    /**
-     * @return array
-     */
-    protected function getTypes()
+    protected function getTypes(): array
     {
         return array_merge(parent::getTypes(), [
             new ProfileFormType('FOS\UserBundle\Tests\TestUser'),