+++ Tests/Security/EmailUserProviderTest.php
@@ -28,7 +28,7 @@ class EmailUserProviderTest extends TestCase
      */
     private $userProvider;
 
-    protected function setUp()
+    protected function setUp(): void
     {
         $this->userManager = $this->getMockBuilder('FOS\UserBundle\Model\UserManagerInterface')->getMock();
         $this->userProvider = new EmailUserProvider($this->userManager);