+++ Tests/Util/CanonicalFieldsUpdaterTest.php
@@ -24,7 +24,7 @@ class CanonicalFieldsUpdaterTest extends TestCase
     private $usernameCanonicalizer;
     private $emailCanonicalizer;
 
-    protected function setUp()
+    protected function setUp(): void
     {
         $this->usernameCanonicalizer = $this->getMockCanonicalizer();
         $this->emailCanonicalizer = $this->getMockCanonicalizer();