+++ Changelog.md
@@ -6,6 +6,8 @@ Changelog
 * [BC break] Change the base class of controllers to use the `AbstractController` (but extending the controllers is not officially supported anymore).
 * [BC break] Remove the group feature
 * [BC break] Change the base class for events to `Symfony\Contracts\EventDispatcher\Event` instead of `Symfony\Component\EventDispatcher\Event`
+* [BC break] Remove the `Symfony\Component\Security\Core\User\AdvancedUserInterface` methods from our `UserInterface`
+* [BC break] Made `\FOS\UserBundle\Model\User::serialize` and `\FOS\UserBundle\Model\User::unserialize` final. Child classes needing to extend the serialization must override `__serialize` and `__unserialize` instead.
 
 ### 2.2.0 (2021-08-26)